apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
itsdangerous==2.2.0
openai-agents==0.2.10
pyjwt==2.10.1
